[refactor]: Page Prefix 추가 및 파일 경로 수정
