+++ src/pages/popup/pages/ScenarioFunnel/pages/ApiSearchPage/apiSearch.css.ts
