+++ src/pages/popup/pages/ScenarioFunnel/pages/Scenario/scenario.css.ts
