+++ src/pages/popup/pages/SequenceFunnel/SequenceFunnel.tsx
@@ -5,9 +5,9 @@ import { useFunnel } from "@/shared/hooks/funnel/useFunnel";
 import { navigationPath } from "@/shared/hooks/useRouter";
 import { useHandleApiList } from "@/widgets/api-list/module/hooks/useHandleApiList";
 import { Dispatch, SetStateAction, useEffect, useState } from "react";
-import DeleteSequencePage from "../Delete/DeleteSequencePage";
-import ReorderSequencePage from "../Reorder/ReorderSequencePage";
-import SequencePage from "../Sequence/SequencePage";
+import { DeleteSequencePage } from "./pages/DeleteSequencePage/DeleteSequencePage";
+import { ReorderSequencePage } from "./pages/ReorderSequencePage/ReorderSequencePage";
+import { SequencePage } from "./pages/SequencePage/SequencePage";
 
 export interface SequenceFunnelProps {
   swaggerTitle: string;
@@ -19,7 +19,7 @@ export interface SequenceFunnelProps {
 // NOTE : 시퀀스 페이지
 // 스퀀스 : 사용자의 API 목록을 담을 수 있는 목록
 
-const SequenceFunnel = () => {
+export const SequenceFunnel = () => {
   const [Funnel, setStep] = useFunnel(
     extractNonEmptyArrayKeys(navigationPath.시나리오_관리_퍼널())
   );
@@ -74,5 +74,3 @@ const SequenceFunnel = () => {
     </Funnel>
   );
 };
-
-export default SequenceFunnel;
