+++ src/pages/popup/pages/SequenceFunnel/pages/SequencePage/module/hooks/useHandleSequencePage.ts
