+++ src/pages/popup/pages/SequenceFunnel/pages/SequencePage/sequence.css.ts
